initial layer-load fade-in
